feat(virtiofs): discover shared memory cache regions
Add transport-level shared-memory descriptors and parse virtio PCI cap64 regions with Linux-compatible ID and range validation. Keep unsupported DAX mount modes disabled while exposing the cache window metadata to virtiofs instances.

Teach PCI BAR discovery to preserve full 64-bit resource sizes, disable address decoding while probing, and leave dedicated shared-memory BARs unmapped so large DAX windows do not consume the MMIO virtual-address pool.

+++ kernel/src/driver/net/e1000e/e1000e.rs
@@ -205,7 +205,7 @@ impl E1000EDevice {
         if address == 0 {
             return Err(E1000EPciError::BarNotAllocated);
         }
-        if size != E1000E_BAR_REG_SIZE {
+        if size != u64::from(E1000E_BAR_REG_SIZE) {
             return Err(E1000EPciError::UnexpectedBarSize);
         }
         let vaddress = bar0
